refactor: separate polaroid list and detail modal
