+++ src/app/board/[boardId]/_components/PolaroidDetailList.tsx
@@ -1,16 +1,16 @@
 'use client'
 
+import React, { useState } from 'react'
 import PolaroidDetailModal from '@/components/Polaroid/PolaroidDetail'
 import { Board } from '@/types'
-import { useState } from 'react'
-import PolaroidListItem from './PolaroidListItem'
+import PolaroidList from '@/components/PolaroidList'
 
-interface PolaroidListProps {
+interface PolaroidDetailListProps {
   board: Board
   boardId: string
 }
 
-const PolaroidList = ({ boardId, board }: PolaroidListProps) => {
+const PolaroidDetailList = ({ board, boardId }: PolaroidDetailListProps) => {
   const [isModalOpen, setIsModalOpen] = useState(false)
   const [selectedIdx, setSelectedIdx] = useState<number>(0)
 
@@ -20,16 +20,8 @@ const PolaroidList = ({ boardId, board }: PolaroidListProps) => {
   }
 
   return (
-    <div className="mx-auto w-full flex-1 overflow-x-hidden overflow-y-scroll pb-10 scrollbar-hide">
-      <div className="grid grid-cols-2 gap-6 px-[20px] py-[10px]">
-        {board.items.map((item, idx) => (
-          <PolaroidListItem
-            key={item.id}
-            item={item}
-            onClick={() => openDetailModal(idx)}
-          />
-        ))}
-      </div>
+    <div>
+      <PolaroidList board={board} onSelectPolaroid={openDetailModal} />
       <PolaroidDetailModal
         isOpen={isModalOpen}
         onClose={() => setIsModalOpen(false)}
@@ -42,4 +34,4 @@ const PolaroidList = ({ boardId, board }: PolaroidListProps) => {
   )
 }
 
-export default PolaroidList
+export default PolaroidDetailList
