+++ src/components/PolaroidList/PolaroidListItem.tsx
